# movies_auth/src/openapi.yaml
openapi: '3.0.0'
info:
  title: Auth OpenAPI
  description: OpenAPI schema for authenticate service of online movies
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
  - url: http://{host}:{port}/api/{version}/auth
    description: Auth server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '5000'
          - '80'
        default: '5000'
      version:
        default: v1
paths:
...
  /check_permission:
    post:
      description: Returns result of permission check
      requestBody:
        description: permission check request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - user_id
                - permission_name
              properties:
                user_id:
                  type: string
                permission_name:
                  type: string
      responses:
        '200':
          description: result of permission check
          content:
            'application/json':
              schema:
                type: object
                required:
                  - result
                properties:
                  result:
                    type: boolean
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
  /login:
    post:
      description: Returns access_token and refresh_token after successful login of user
      requestBody:
        description: login request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: successful user login
          content:
            'application/json':
              schema:
                type: object
                required:
                  - access_token
                  - refresh_token
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
  /logout:
    get:
      description: Logout user
      responses:
        '204':
          description: successful user logout
          content:
            'text/html':
              schema:
                type: string
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
...
  /permission:
    post:
      description: Returns created permission
      requestBody:
        description: permission request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        '200':
          description: created permission
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Permission'
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
  /permission/{permission_id}:
    put:
      description: Returns changed permission
      requestBody:
        description: permission request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        '200':
          description: changed permission
          content:
            'application/json':
              schema:
                type: object
                required:
                  - item
                  - message
                  - properties
                properties:
                  item:
                    $ref: '#/components/schemas/Permission'
                  message:
                    type: string
                  properties:
                    type: object
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
      - name: permission_id
        in: path
        description: permission uuid
        required: true
        schema:
          type: string
...
  /role_permission_add:
    post:
      description: Assigns permission to role
      requestBody:
        description: role permission request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - role_id
                - permission_id
              properties:
                role_id:
                  type: string
                permission_id:
                  type: string
      responses:
        '200':
          description: permission assigned to role
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/RolePermission'
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
  /role_permission_revoke:
    post:
      description: Revokes permission from role
      requestBody:
        description: role permission request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - role_id
                - permission_id
              properties:
                role_id:
                  type: string
                permission_id:
                  type: string
      responses:
        '200':
          description: permission revoked from role
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/RolePermission'
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    Permission:
      type: object
      required:
        - uuid
        - name
      properties:
        uuid:
          type: string
        name:
          type: string
...
    RolePermission:
      type: object
      required:
        - uuid
        - role_id
        - permission_id
      properties:
        uuid:
          type: string
        role_id:
          type: string
        permission_id:
          type: string
...